requirements.txt
Flask
gunicorn 
google-cloud-logging